# ai/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          },
          "requestId": {
            "format": "uuid",
            "title": "Requestid",
            "type": "string"
          }
        },
        "required": [
          "error",
          "requestId"
        ],
        "title": "ApiErrorResponse",
        "type": "object"
      },
...
      "EntityProfileRequest": {
        "additionalProperties": false,
        "properties": {
          "contractVersion": {
            "const": "v1",
            "title": "Contractversion",
            "type": "string"
          },
          "entity": {
            "$ref": "#/components/schemas/EntityProfileInput"
          },
          "inputFingerprint": {
            "pattern": "^[a-fA-F0-9]{64}$",
            "title": "Inputfingerprint",
            "type": "string"
          },
          "requestId": {
            "format": "uuid",
            "title": "Requestid",
            "type": "string"
          },
          "tenantId": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Tenantid",
            "type": "string"
          }
        },
        "required": [
          "requestId",
          "contractVersion",
          "tenantId",
          "inputFingerprint",
          "entity"
        ],
        "title": "EntityProfileRequest",
        "type": "object"
      },
      "EntityProfileResult": {
        "additionalProperties": false,
        "properties": {
          "capabilities": {
            "items": {
              "type": "string"
            },
            "maxItems": 200,
            "title": "Capabilities",
            "type": "array"
          },
          "components": {
            "items": {
              "type": "string"
            },
            "maxItems": 200,
            "title": "Components",
            "type": "array"
          },
          "coverage": {
            "items": {
              "$ref": "#/components/schemas/ProfileCoverage"
            },
            "maxItems": 500,
            "title": "Coverage",
            "type": "array"
          },
          "entityId": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Entityid",
            "type": "string"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ServiceError"
            },
            "title": "Errors",
            "type": "array"
          },
          "failureModes": {
            "items": {
              "type": "string"
            },
            "maxItems": 200,
            "title": "Failuremodes",
            "type": "array"
          },
          "generatedDescription": {
            "maxLength": 20000,
            "minLength": 1,
            "title": "Generateddescription",
            "type": "string"
          },
          "profileFingerprint": {
            "pattern": "^[a-fA-F0-9]{64}$",
            "title": "Profilefingerprint",
            "type": "string"
          },
          "profileId": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Profileid",
            "type": "string"
          },
          "profileVersion": {
            "minimum": 1.0,
            "title": "Profileversion",
            "type": "integer"
          },
          "provenance": {
            "$ref": "#/components/schemas/ProfileProvenance"
          },
          "requestId": {
            "format": "uuid",
            "title": "Requestid",
            "type": "string"
          },
          "searchHints": {
            "items": {
              "type": "string"
            },
            "maxItems": 200,
            "title": "Searchhints",
            "type": "array"
          },
          "status": {
            "enum": [
              "upserted",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "systems": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "title": "Systems",
            "type": "array"
          }
        },
        "required": [
          "requestId",
          "status",
          "entityId",
          "profileVersion",
          "profileId",
          "generatedDescription",
          "profileFingerprint",
          "provenance"
        ],
        "title": "EntityProfileResult",
        "type": "object"
      },
...
      "ExtractRequest": {
        "additionalProperties": false,
        "properties": {
          "contractVersion": {
            "const": "v1",
            "title": "Contractversion",
            "type": "string"
          },
          "declaredMetadata": {
            "$ref": "#/components/schemas/DeclaredMetadata"
          },
          "documentId": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Documentid",
            "type": "string"
          },
          "documentVersionId": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Documentversionid",
            "type": "string"
          },
          "requestId": {
            "format": "uuid",
            "title": "Requestid",
            "type": "string"
          },
          "sourceFile": {
            "$ref": "#/components/schemas/SourceFile"
          },
          "versionNumber": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Versionnumber",
            "type": "string"
          }
        },
        "required": [
          "requestId",
          "contractVersion",
          "documentId",
          "documentVersionId",
          "versionNumber",
          "sourceFile",
          "declaredMetadata"
        ],
        "title": "ExtractRequest",
        "type": "object"
      },
      "ExtractResult": {
        "additionalProperties": false,
        "properties": {
          "documentSummary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DocumentSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "documentVersionId": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Documentversionid",
            "type": "string"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ServiceError"
            },
            "title": "Errors",
            "type": "array"
          },
          "extractedMetadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExtractedMetadata"
              },
              {
                "type": "null"
              }
            ]
          },
          "extractionQuality": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Extractionquality",
            "type": "number"
          },
          "requestId": {
            "format": "uuid",
            "title": "Requestid",
            "type": "string"
          },
          "status": {
            "enum": [
              "needs_review",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "requestId",
          "status",
          "documentVersionId",
          "extractionQuality"
        ],
        "title": "ExtractResult",
        "type": "object"
      },
...
      "IndexRequest": {
        "additionalProperties": false,
        "properties": {
          "approvalState": {
            "const": "APPROVED",
            "title": "Approvalstate",
            "type": "string"
          },
          "contractVersion": {
            "const": "v1",
            "title": "Contractversion",
            "type": "string"
          },
          "documentId": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Documentid",
            "type": "string"
          },
          "documentVersionId": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Documentversionid",
            "type": "string"
          },
          "requestId": {
            "format": "uuid",
            "title": "Requestid",
            "type": "string"
          },
          "reviewedMetadata": {
            "$ref": "#/components/schemas/ReviewedMetadata"
          },
          "sourceFile": {
            "$ref": "#/components/schemas/SourceFile"
          }
        },
        "required": [
          "requestId",
          "contractVersion",
          "documentId",
          "documentVersionId",
          "approvalState",
          "reviewedMetadata",
          "sourceFile"
        ],
        "title": "IndexRequest",
        "type": "object"
      },
      "IndexResult": {
        "additionalProperties": false,
        "properties": {
          "chunkCount": {
            "minimum": 0.0,
            "title": "Chunkcount",
            "type": "integer"
          },
          "contentFingerprint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contentfingerprint"
          },
          "documentVersionId": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Documentversionid",
            "type": "string"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ServiceError"
            },
            "title": "Errors",
            "type": "array"
          },
          "indexReference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Indexreference"
          },
          "requestId": {
            "format": "uuid",
            "title": "Requestid",
            "type": "string"
          },
          "sourceLocations": {
            "items": {
              "$ref": "#/components/schemas/SourceLocation"
            },
            "title": "Sourcelocations",
            "type": "array"
          },
          "status": {
            "enum": [
              "indexed",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "requestId",
          "status",
          "documentVersionId",
          "chunkCount"
        ],
        "title": "IndexResult",
        "type": "object"
      },
      "LogDraftRequest": {
        "additionalProperties": false,
        "properties": {
          "citationIds": {
            "items": {
              "maxLength": 200,
              "minLength": 1,
              "type": "string"
            },
            "title": "Citationids",
            "type": "array"
          },
          "contractVersion": {
            "const": "v1",
            "title": "Contractversion",
            "type": "string"
          },
          "equipmentId": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Equipmentid"
          },
          "projectId": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Projectid",
            "type": "string"
          },
          "requestId": {
            "format": "uuid",
            "title": "Requestid",
            "type": "string"
          },
          "scopeType": {
            "$ref": "#/components/schemas/LogScope"
          },
          "sourceText": {
            "maxLength": 20000,
            "minLength": 1,
            "title": "Sourcetext",
            "type": "string"
          }
        },
        "required": [
          "requestId",
          "contractVersion",
          "projectId",
          "scopeType",
          "sourceText"
        ],
        "title": "LogDraftRequest",
        "type": "object"
      },
      "LogDraftResult": {
        "additionalProperties": false,
        "properties": {
          "citationIds": {
            "items": {
              "maxLength": 200,
              "minLength": 1,
              "type": "string"
            },
            "title": "Citationids",
            "type": "array"
          },
          "draftText": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Drafttext"
          },
          "requestId": {
            "format": "uuid",
            "title": "Requestid",
            "type": "string"
          },
          "status": {
            "enum": [
              "drafted",
              "unavailable"
            ],
            "title": "Status",
            "type": "string"
          },
          "suggestedFields": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Suggestedfields",
            "type": "object"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "requestId",
          "status"
        ],
        "title": "LogDraftResult",
        "type": "object"
      },
...
      "ReadinessStatus": {
        "additionalProperties": false,
        "properties": {
          "service": {
            "const": "patch-ai",
            "default": "patch-ai",
            "title": "Service",
            "type": "string"
          },
          "status": {
            "enum": [
              "ready",
              "unavailable"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "ReadinessStatus",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Private ingestion, retrieval, and generation service for P.A.T.C.H.",
    "title": "P.A.T.C.H. AI Service",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/readiness": {
      "get": {
        "description": "Aggregate runtime readiness without exposing configuration names or values.",
        "operationId": "readiness_readiness_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadinessStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request correlation mismatch."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service authentication failed."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request ID replay detected."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request schema validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unexpected internal failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service authentication unavailable."
          }
        },
        "summary": "Readiness",
        "tags": [
          "system"
        ]
      }
    },
    "/v1/entity-profiles/upsert": {
      "post": {
        "operationId": "upsert_entity_profile_v1_entity_profiles_upsert_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityProfileResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request correlation mismatch."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service authentication failed."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request ID replay detected."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request schema validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unexpected internal failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service authentication unavailable."
          }
        },
        "summary": "Upsert Entity Profile",
        "tags": [
          "entity-profiles"
        ]
      }
    },
    "/v1/ingestions/extract": {
      "post": {
        "operationId": "extract_document_v1_ingestions_extract_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request correlation mismatch."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service authentication failed."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request ID replay detected."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request schema validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unexpected internal failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service authentication unavailable."
          }
        },
        "summary": "Extract Document",
        "tags": [
          "ingestion"
        ]
      }
    },
    "/v1/ingestions/index": {
      "post": {
        "operationId": "index_document_v1_ingestions_index_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IndexRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IndexResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request correlation mismatch."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service authentication failed."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request ID replay detected."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request schema validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unexpected internal failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service authentication unavailable."
          }
        },
        "summary": "Index Document",
        "tags": [
          "ingestion"
        ]
      }
    },
    "/v1/log-drafts": {
      "post": {
        "operationId": "draft_log_v1_log_drafts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogDraftResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request correlation mismatch."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service authentication failed."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request ID replay detected."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request schema validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unexpected internal failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service authentication unavailable."
          }
        },
        "summary": "Draft Log",
        "tags": [
          "drafts"
        ]
      }
    },
    "/v1/procedure-drafts": {
      "post": {
        "operationId": "draft_procedure_v1_procedure_drafts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProcedureDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProcedureDraftResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request correlation mismatch."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service authentication failed."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request ID replay detected."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request schema validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unexpected internal failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service authentication unavailable."
          }
        },
        "summary": "Draft Procedure",
        "tags": [
          "drafts"
        ]
      }
    },
    "/v1/questions": {
      "post": {
        "operationId": "answer_question_v1_questions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request correlation mismatch."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service authentication failed."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request ID replay detected."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Request schema validation failed."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unexpected internal failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service authentication unavailable."
          }
        },
        "summary": "Answer Question",
        "tags": [
          "questions"
        ]
      }
    }
  }
}